[qtc] config file rewriting and alt src directory setting
* Add SDK install directory path to mersdk.xml configuration
file. If install directory change is detected when qtcreator
starts up, it will write new settings from the new mersdk.xml
and deregister all old targets (kits).

TODO: the old target specific scripts should be removed from the
$HOME/.config/... path when the kits are deregistered. It is not
a big issue to leave them there.

* Alternative src directory can be updated from Mer settings page.
This requires that MerSDK is not running when the folder is changed.

* Add more tooltips to the start/stop VM buttons when the startup or
shutdown is proceeding.

Signed-off-by: Juha Kallioinen <juha.kallioinen@jolla.com>