Mer hack - use install dir as the user settings dir for windows
The problem in windows with the default user settings path, which
is %APPDATA%/SettingsVariant comes when the windows user's username
contains non-ascii characters.

The %APPDATA% expands to something like (in Windows 7):

C:\users\John Döe\AppData\Roaming\

This path ends up in the batch files for qmake/gcc/deploy etc. and
batch files do not handle non-ascii very well. Trying to create
the batch files with additional code page changes to "chcp 65001"
was not enough to make it work.

Using the install dir as the settings dir is a workaround that works
as long as the user has write access to the install dir, which he
should have since he installed it there.

Signed-off-by: Juha Kallioinen <juha.kallioinen@jolla.com>
(cherry picked from commit 989f332666ea67456643a72c201877753561b212)