Reintroduce "Jolla: Remove wizards which do not properly work with SailfishOS"
This reintroduces commit 6e20d1f9d6cbbf01570ae58c18b2610970443e62,
temporarily reverted prior to mer-3.1 merge with commit
7716dd79e488228556655276fe646383f3f0ec03.

BlackBerry wizard display name added because the very first wizard in a
category determines the display name for the category - it would be
empty. This is not an ideal solution as the display name does not get
localized this way. It would not be trivial to fix this properly.

Signed-off-by: Martin Kampas <martin.kampas@jolla.com>