Mer: Add RPM validation deploy step
The "RPM Validation" step is implicit to the "Deploy By Building An RPM
Package" method and by default it is enabled initially - this can be
changed in options.

Attention: To try this change SDK virtual machine must be unset and
re-set in qtc options. Also existing project configuration is not
touched, so projects must be reconfigured.

Signed-off-by: Martin Kampas <martin.kampas@jolla.com>