Mer: Add VM resizing to SDK details widget
Added a new spin box labeled “Disk image capacity” to
`MerSdkDetailsWidget` for showing and setting the maximum capacity of
the disk image file associated with a virtual machine.

Initially the size shown is the current capacity of the disk image. To
resize the disk image, a capacity greater than the current one has to
be entered. Pressing the “Resize” button will show a warning message if
the capacity is invalid.

In `MerOptionsWidget`, the new slot `onResizeDiskImageButtonClicked()`
takes care of closing the virtual machine for the duration of the
resizing operation and restarting it, if necessary.

The existing slot `onSrcFolderApplyButtonClicked()` was refactored to
use the new private utility methods `showCloseVirtualMachineDialog()`
and `showStartVirtualMachineDialog()`, as the functionality of these
two slots is very similar.

Note that even though the functionality introduced here correctly
increases the maximum capacity of the disk image, the extra capacity is
not taken into use until corresponding resizing support is introduced
to the build engine virtual machine.

For this reason it is possible to disable the resizing feature by
defining the flag `MER_NO_BUILD_ENGINE_RESIZE` when building Qt
Creator. If this flag is defined, the capacity of the disk image is
shown in the SDK details but it is not possible to modify it.