Mer: Disable VM disk image resizing if capacity is unchanged
The “Resize” button of the SDK details widget is now disabled if the
entered disk image capacity is not greater than the current capacity.

The redundant warning dialog related to the situation where the current
and new capacities are identical was removed.

The “Resize” button is now also disabled while resizing is ongoing.
This fixes an issue where it was possible to click the “Resize” button
multiple times resulting is confusing error/success dialogs.