Mer: Make web UI port editable in SDK details widget (Fixes JB#30222)
Replaced the static web UI port label with a spin box in
`MerSdkDetailsWidget` to make it possible for the user to change the
port of the build engine web UI.

The initial value for the spin box is the port configured by the
installer (8080). The user can specify any port in the range
[1024..65535].

No attempt is made to verify that the port is actually free. The SDK
configuration and the NAT port forwarding rule of the “MerSDK” virtual
machine is updated as soon as the user pressed the “Change” button.

In `MerOptionsWidget`, the new slot `onWwwPortApplyButtonClicked(int)`
takes care of closing the virtual machine for the duration of the port
update operation and restarting it, if necessary.

The existing slot `onSrcFolderApplyButtonClicked()` was refactored to
use the new private utility methods `showCloseVirtualMachineDialog()`
and `showStartVirtualMachineDialog()`, as the functionality of these
two slots is very similar.