Mer: Do not use device type ID to distinguish architecture
With the original approach the following terms are used interchangeably:

    Mer i486 device <--> Mer emulator device
        (var Constants::MER_DEVICE_TYPE_I486, class MerEmulatorDevice)
    Mer ARM device  <--> Mer hardware device
        (var Constants::MER_DEVICE_TYPE_ARM, class MerHardwareDevice)

The distinction on the left side (i486/ARM) expresses tool chain
compatibility with the device - presented to the user when associating a
device with a kit.

The distinction on the right side (emulator/hardware) expresses the access
method used to communicate with the device - presented to the user when
creating a device.

The meaning of the terms on the left side compared to the right side is
threfore very different and using these interchangeably makes no sence.
This appeared obvious after the introduction of tablet (physical)
device, which is x86.

This commit allows any combination of the options from both sides by
replacing the two recognized IDevice::type values MER_DEVICE_TYPE_I486 &
MER_DEVICE_TYPE_ARM with single value MER_DEVICE_TYPE, and storing the
information as MerDevice::architecture instead.

The drawback is that it is now possible to set an x86 device to an ARM
kit (and vice versa) and this is only detected during kit validation.
This was chosen as the less bad option, because IDevice::type is also
used by device factory and possible values are presented to the user to
choose from when creating a device (the user would be forced to choose
if he wants to create an x86 or an ARM device; now he is asked if an
emulator or a physical device should be created.)