Mer: Wrap MerEmulatorStartStep in MerPrepareTargetStep
MerEmulatorStartStep is specific to deployment to an emulator device.
Including this step _conditionally_ in a deploy configuration requires
Kit::device be set at the time of creation of the deploy configuration
(to be able to check IDevice::machineType).

This commit wraps MerEmulatorStartStep in MerPrepareTargetStep.
MerPrepareTargetStep chooses either MerEmulatorStartStep (for emulator
device) or the newly added MerConnectionTestStep (for physical device)
as a "strategy" later upon init() call.

MerPrepareTargetStep is always present in the deploy configuration where
MerEmulatorStartStep would occur conditionally before. This removes the
need for Kit::device be set at the time a deploy configuration is being
created.