Mer: Wait for gdbserver to become ready
This fixes random failures upon start of debugging session with errors
like "putpkt: write failed: Broken pipe." or "Remote communication
error. Target disconected." (depending on GDB version).

There is a race condition when gdbserver is started in parallel with
GDB. Normally GDB can deal with this - see the documentation of GDB's
options "tcp auto-retry" and "tcp connect-timeout". Unfortunately GDB's
auto-retry mechanism does not work with various forwarded connections
which is the case of e.g. VirtualBox port forwarding and SSH port
forwarding.

Most likely there used to be some delay in the process of starting GDB
and so this was not an issue up to QtC 4.2.  I haven't found any obvious
change between 4.2 and 4.4 codebase that could trigger this behavior.

As a workaround GdbServerReadyWatcher delays starting of GDB until
gdbserver opens its TCP port.