Revert "qmake: require a drive in a DOS path for it to be absolute"
This reverts commit 912dd8df64c05fcdcbc9d862a41052654dd07546.

The commit broke code completion on Windows hosts because it does not
consider path conventions on the target platform. Here the target
platform is Unix-like. Having an absolute path specified in a project
file, e.g. with INCLUDEPATHS += /usr/include/foo, it would be considered
as relative path on Windows hosts and prefixed with build directory path
as a fallback method to get an absolute path from it. Instead it should
be treated as an absolute path, in which case it would be later prefixed
with sysroot path, which is the desired behavior.

(cherry picked from commit 71ea6aec9b58ea099fae9e0a6f10ccdab6cb0a12)