Sfdk: Allow to run engine command with ForwardedInputChannel
This enables more terminal features like turning off echo for reading
passwords.

OTOH passing too much information about the terminal to the remote
application may not be desired. Consider an application animating a
progress bar up to the right edge of the terminal. If that progress bar
follows a file name that is subject to reverse path mapping, it breaks
easily as the paths on host are usualy longer than the paths under the
build engine.  For that reason it is disabled by default.