Mer: Work around VM info not immediately available.
VM info initialization is only started during Sdk::enableUpdates and it
is done asynchronously, so it still takes some time after
Sdk::enableUpdates returns before the info becomes available.  It is not
clear yet how to fix it properly and at the same time it is known to
only become an issue when Options dialog is opened very soon (<3 secs)
after application startup, hence this workaround.