Sfdk: Notify other processes about VM changes
Some of the VM properties are persisted by BuildEngine and Emulator
classes, some not. Those that are persisted are already immediately
notified about thanks to the shared settings mechanism. Those that are
NOT persisted need to be notified some other way. This commit enables
use of the VM info cache file for that purpose. The VM info cache is
updated on every change to VM configuration and so watching the cache
file is a reliable way to get notified about all VM changes done by
other libsfdk-based processes.