Mer: Warn in .spec/.yaml file is not registered to project
Only those .spec/.yaml files that are registered to the project are
considered for triggering qmake execution in order to augment project
model using qmake variables added by rpmbuild. (Because we need to run
qmake in order to update those, and this is only possible with the
active project.)

Warn the user if the file is not registered to the project but avoid
being too specific here as it is not easy to explain in a single
sentence what is it needed for.