Qmake: Obey QmakeStep::extraParserArguments when comparing to Makefile
If additional arguments are added on qmake command line lower in the
stack, they appear in the Makefile and comparison would fail, leading to
the "The qmake arguments have changed. The build will be overwritten"
warning under the "Build directory" line edit in Build Settings. These
additional qmake arguments are already accessible to the project parser
via QmakeStep::extraParserArguments().