CMake: Do not reparse project on changes during build
We need to maintain two versions of CMakeCache.txt and *.cbp files - one
for Qt Creator (with paths mapped) and one for CMake itself, which is
running under the build engine. And we swap those two versions before
and after build. Qt Creator started noticing the change after upgrade to
4.12.2.

This leands to a temporary state when the project is not properly parsed
during build, warnings appear that files are not part of any project and
debugger may receive wrong paths to name some of the bad consequences.

Suppressing it this way is potentionally dangerous as an actual change
might get ignored, but no such issue could be reproduced so far.

This commit is not suitable for upstream.

(cherry picked from commit e4871ef1787cbb7ea7047a8d4fdf840583a36009)