Sfdk: Drop option to use managed input for remote commands
The reason why this feature existed is that before, when fixed mount
locations were used for the shared workspace directory(ies), reverse
mapping in command output usually changed lengths of strings, which is
problematic with screen-oriented programs.  Use of ManagedInputChannel
prevents the remote process to see a terminal is attached and so if
there are optional screen-oriented functions it does not enable them.

This was later found to be broken on Windows (or just on more recent
Windows versions).

Now that the aligned mount point paths are used, this workaround is not
needed anymore.