MerSsh: Fix unquoting arguments
This fixes commit bcf7d8ba2b88b3c6039be641cd73b1879955f9a7 (MerSsh:
Always use sfdk to invoke remote commands).

It seems no one ever really understood how the unquoting worked.

Actually unquoteArguments never had any effect. Even the part that was
supposed to check that arguments are quoted when needed never worked.

When merssh is invoked through a wrapper and the wrapper is passed
single argument "Foo Bar Baz", it invokes merssh with three arguments
"'Foo", "Bar" and "Baz'". So merssh may never receive an argument with
spaces embedded, quoted or unquoted. (No, this cannot be done a better
way.  This is the least bad method given the limitations of CMD on
Windows AFAIK.)

So "'Foo", "Bar" and "Baz'" were passed through untouched by
unquoteArguments and forwarded to ssh, which (to many people's surprise)
passes all its nonoption arguments joined with spaces as a single
argument to "sh -c" on the remote side.

Thanks to this (ugly) behavior of ssh this used to work. When sfdk was
taken into use, which does perfect argument forwarding, it broke.