Sfdk: Do not immediately hard-close VM to stop it
Sfdk, in order to minimize startup time, makes some assumptions and does
not fully synchronize VM status unless really required. As a result the
VM is in a zombie state most of the sfdk execution time. VmConnection
OTOH does not try to soft-close VM if it was not in the connected state
before.