Sfdk: Revert emulator configuration version to 1
Should we take it seriously, we would need to maintain separate versions
for each emulator configuration and the container. Upstream does not do
anything like that. Their configuration files are all at version 1
despite the fact that they have clearly undergone some evolution, so
they handle it elsewhere.

In our case it is enough this time that the format of emulator IDs has
changed, so we do not need to do anything special and the obsolete
configuration will be simply dropped.