Sfdk: Ensure stable order of emulators
Main motivation for doing this is that the QtC device created for the
very first emulator becomes the default device of the given type, and
there is no API to override. So it is important that the latest emulator
comes first.

Doing similarly for devices ensures UI consistency and stable results
for the very first and subsequent executions.