Mer: Fix false complains on debug session in progress at shutdown
The original behavior does not use gdbserver with --once and instead use
"monitor exit" prior to closing the local GDB. This leads to "Remote
connection closed" error which the debugger engine state machine does
not expect and ends in some intermediate state.

Note that it is still possible to see it complaining on stderr that GDB
WON'T EXIT when debugging session is closed when application is stopped
on an QML breakpoint. But this is handled well by the state machine.