Sfdk: Utilize man (the manual pager) for documentation
A minimal AsciiDoc-like format is now used as an intermediate format for
documentation, which can be either translatted to pretty formatted man
pages with asciidoc (a2x more exactly) or displayed with simple
formatting directly by sfdk.

Manual pages should be generated at build time and packaged with Qt
Creator. Sfdk first tries to display a man page, falling back to its own
simple rendering capability when the man page or the man pager is not
available.

This required reorganization of the documentation: commands help is no
more included in the --help-* manuals.