Sfdk: Avoid noncritical device connection test
Save a little bit of time.  When invoked by Qt Creator, connection was
already tested prior to this point. What we really need to do here is to
run the possible user hook and autostart the emulator, and for that it
is enough to prepare the remote process, no need to actually run it.