Sfdk: Bind mount /dev under Docker container on Linux
Main motivation here is to get access to dynamically created loop and
LVM devices, which are needed for device image creation. It still needs
the '--privileged' option to actually allow accessing these devices
(controlled by the SAILFISH_SDK_DOCKER_RUN_PRIVILEGED env var). Or from
the other side: with the defauts it is hardly possible to screw
something up on the host through the bind-mounted /dev.