Dist/GDB: Use Python 3.10
The main motivation for upgrading is that the prebuilt Python binary we
have been using on Windows is 32 bit, so needs to be rebuilt. At the
same time the old Python does not build under MinGW-w64 because of
https://github.com/python/cpython/issues/62131 (at least).

Windows-related notes follows.

There is no 3.10.14 branch in cpython-mingw, hence the use of 3.10.12
on Windows.

It is inevitable to use --prefix=/ucrt64 when building on Windows, hence
the final copy to the desired location with the tar pipe. Sole
DESTDIR use would include the prefix, which is unwanted.

It must be explicitly told the library directory path with LDFLAGS. It
can figure the include path itself, but fails on the library path.